리스트 카드 모듈화
